@@ -21,39 +21,39 @@
   },
   "homepage": "https://github.com/digitalbazaar/bedrock-vc-verifier",
   "dependencies": {
-    "@digitalbazaar/ed25519-signature-2018": "^2.1.0",
-    "@digitalbazaar/ed25519-signature-2020": "^3.0.0",
-    "@digitalbazaar/vc": "^2.1.0",
-    "@digitalbazaar/vc-status-list": "^4.0.0",
+    "@digitalbazaar/ed25519-signature-2018": "^3.0.0",
+    "@digitalbazaar/ed25519-signature-2020": "^4.0.1",
+    "@digitalbazaar/vc": "^3.0.0",
+    "@digitalbazaar/vc-revocation-list": "^4.0.0",
+    "@digitalbazaar/vc-status-list": "^5.0.0",
     "assert-plus": "^1.0.0",
-    "bnid": "^2.1.0",
-    "body-parser": "^1.19.0",
-    "cors": "^2.8.5",
-    "vc-revocation-list": "^3.0.0"
+    "bnid": "^3.0.0",
+    "body-parser": "^1.20.0",
+    "cors": "^2.8.5"
   },
   "peerDependencies": {
-    "@bedrock/core": "^6.0.0",
+    "@bedrock/core": "^6.0.1",
     "@bedrock/credentials-context": "^3.0.0",
     "@bedrock/did-context": "^4.0.0",
-    "@bedrock/did-io": "^8.0.0",
+    "@bedrock/did-io": "^9.0.1",
     "@bedrock/express": "^8.0.0",
     "@bedrock/https-agent": "^4.0.0",
     "@bedrock/jsonld-document-loader": "^3.0.0",
     "@bedrock/mongodb": "^10.0.0",
     "@bedrock/security-context": "^7.0.0",
-    "@bedrock/service-agent": "^5.0.0",
-    "@bedrock/service-context-store": "^7.0.0",
-    "@bedrock/service-core": "^5.0.0",
-    "@bedrock/vc-status-list-context": "^4.0.0",
-    "@bedrock/vc-revocation-list-context": "^3.0.0",
-    "@bedrock/veres-one-context": "^14.0.0",
-    "@bedrock/validation": "^7.0.0"
+    "@bedrock/service-agent": "^6.0.0",
+    "@bedrock/service-context-store": "^8.0.0",
+    "@bedrock/service-core": "^6.0.0",
+    "@bedrock/validation": "^7.0.0",
+    "@bedrock/vc-revocation-list-context": "^3.1.0",
+    "@bedrock/vc-status-list-context": "^4.1.0",
+    "@bedrock/veres-one-context": "^14.0.1"
   },
   "devDependencies": {
-    "eslint": "^7.32.0",
-    "eslint-config-digitalbazaar": "^2.8.0",
-    "eslint-plugin-jsdoc": "^28.6.1",
-    "jsdoc": "^3.6.4",
+    "eslint": "^8.18.0",
+    "eslint-config-digitalbazaar": "^3.0.0",
+    "eslint-plugin-jsdoc": "^39.3.3",
+    "jsdoc": "^3.6.10",
     "jsdoc-to-markdown": "^7.1.1"
   },
   "engines": {
