@@ -4,9 +4,7 @@
 import * as bedrock from '@bedrock/core';
 import * as database from '@bedrock/mongodb';
 import assert from 'assert-plus';
-import {createRequire} from 'node:module';
-const require = createRequire(import.meta.url);
-const {generateId, decodeId} = require('bnid');
+import {generateId, decodeId} from 'bnid';
 
 const {util: {BedrockError}} = bedrock;
 
