@@ -2,6 +2,7 @@
  * Copyright (c) 2018-2022 Digital Bazaar, Inc. All rights reserved.
  */
 import * as bedrock from '@bedrock/core';
+import * as vc from '@digitalbazaar/vc';
 import {createChallenge, verifyChallenge} from './challenges.js';
 import {asyncHandler} from '@bedrock/express';
 import bodyParser from 'body-parser';
@@ -13,15 +14,10 @@ import {
   verifyPresentationBody
 } from '../schemas/bedrock-vc-verifier.js';
 import {createDocumentLoader} from './documentLoader.js';
-import {createRequire} from 'node:module';
 import {createValidateMiddleware as validate} from '@bedrock/validation';
 import {metering, middleware} from '@bedrock/service-core';
-const require = createRequire(import.meta.url);
-const {Ed25519Signature2020} =
-  require('@digitalbazaar/ed25519-signature-2020');
-const {Ed25519Signature2018} =
-  require('@digitalbazaar/ed25519-signature-2018');
-const vc = require('@digitalbazaar/vc');
+import {Ed25519Signature2018} from '@digitalbazaar/ed25519-signature-2018';
+import {Ed25519Signature2020} from '@digitalbazaar/ed25519-signature-2020';
 
 const {util: {BedrockError}} = bedrock;
 
