@@ -2,16 +2,14 @@
  * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
 import assert from 'assert-plus';
-import {createRequire} from 'node:module';
-const require = createRequire(import.meta.url);
-const {
-  checkStatus: statusListCheckStatus,
-  statusTypeMatches: statusListStatusTypeMatches
-} = require('@digitalbazaar/vc-status-list');
-const {
-  checkStatus: revocationListCheckStatus,
-  statusTypeMatches: revocationListStatusTypeMatches
-} = require('vc-revocation-list');
+import {
+  checkStatus as revocationListCheckStatus,
+  statusTypeMatches as revocationListStatusTypeMatches
+} from '@digitalbazaar/vc-revocation-list';
+import {
+  checkStatus as statusListCheckStatus,
+  statusTypeMatches as statusListStatusTypeMatches
+} from '@digitalbazaar/vc-status-list';
 
 const handlerMap = new Map();
 handlerMap.set('RevocationList2020Status', {
