@@ -3,16 +3,14 @@
  */
 import * as bedrock from '@bedrock/core';
 import {httpsAgent} from '@bedrock/https-agent';
-import {createRequire} from 'module';
 import {didIo} from '@bedrock/did-io';
 import {getAppIdentity} from '@bedrock/app-identity';
 import {mockData} from './mock.data.js';
 import {httpClient} from '@digitalbazaar/http-client';
-const require = createRequire(import.meta.url);
-const {Ed25519Signature2020} = require('@digitalbazaar/ed25519-signature-2020');
-const {EdvClient} = require('@digitalbazaar/edv-client');
-const {KeystoreAgent, KmsClient} = require('@digitalbazaar/webkms-client');
-const {ZcapClient} = require('@digitalbazaar/ezcap');
+import {Ed25519Signature2020} from '@digitalbazaar/ed25519-signature-2020';
+import {EdvClient} from '@digitalbazaar/edv-client';
+import {KeystoreAgent, KmsClient} from '@digitalbazaar/webkms-client';
+import {ZcapClient} from '@digitalbazaar/ezcap';
 
 const edvBaseUrl = `${mockData.baseUrl}/edvs`;
 const kmsBaseUrl = `${mockData.baseUrl}/kms`;
